# controller-model/routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "The APIs for zlz - a photo sharing social network - are listed below for your use. For the source code of our project, please see [https://github.com/cis557/cis557-project-photo-sharing-social-network-app-zlz](https://github.com/cis557/cis557-project-photo-sharing-social-network-app-zlz)",
    "version": "1.0.0",
    "title": "zlz - photo sharing social network",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "zzm@seas.upenn.edu"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "peaceful-forest-83015.herokuapp.com/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/like": {
      "get": {
        "tags": [
          "like"
        ],
        "summary": "READ whether the user has liked the post",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "liker",
            "type": "string",
            "description": "id of the user",
            "required": true
          },
          {
            "in": "query",
            "name": "likee",
            "type": "string",
            "description": "id of the post",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "fail to check whether the user has liked the post"
          }
        }
      }
    },
    "/like/put": {
      "get": {
        "tags": [
          "like"
        ],
        "summary": "UPDATE the liking relationship between a user and a post",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "liker",
            "type": "string",
            "description": "id of the user",
            "required": true
          },
          {
            "in": "body",
            "name": "likee",
            "type": "string",
            "description": "id of the post",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "fail to like/unlike the post"
          }
        }
      }
    },
    "/follow": {
      "get": {
        "tags": [
          "follow"
        ],
        "summary": "READ whether the follower has followed the followee",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "follower",
            "type": "string",
            "description": "id of the follower",
            "required": true
          },
          {
            "in": "query",
            "name": "followee",
            "type": "string",
            "description": "id of the followee",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "missing parameters"
          },
          "404": {
            "description": "fail to check whether the follower has followed the followee"
          }
        }
      }
    },
    "/follow/put": {
      "get": {
        "tags": [
          "follow"
        ],
        "summary": "UPDATE the followship between two users",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "liker",
            "type": "string",
            "description": "id of the follower",
            "required": true
          },
          {
            "in": "body",
            "name": "likee",
            "type": "string",
            "description": "id of the followee",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "missing parameters"
          },
          "404": {
            "description": "fail to like/unlike the post"
          }
        }
      }
    },
    "/comment": {
      "put": {
        "tags": [
          "comment"
        ],
        "summary": "CREATE a comment for a post",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postId",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "authorId",
            "description": "id of the author",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "content",
            "description": "content of the comment",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "mention",
            "description": "mentioned user list",
            "required": true,
            "type": "array"
          }
        ],
        "responses": {
          "201": {
            "description": "comment created"
          },
          "400": {
            "description": "missing information"
          },
          "404": {
            "description": "fail to create the comment"
          }
        }
      }
    },
    "/comment/delete": {
      "put": {
        "tags": [
          "comment"
        ],
        "summary": "DELETE a comment for a post",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postId",
            "description": "id of the post",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "commentId",
            "description": "id of the comment",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "comment deleted"
          },
          "404": {
            "description": "fail to delete the comment"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "user login",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "username",
            "description": "username of the user",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "password",
            "description": "password of the user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "user registration",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "username",
            "description": "username of the user",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "image",
            "description": "image of the user in base64 format",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "password",
            "description": "password of the user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
  }
}